新增seniortable
