+++ .prettierrc
@@ -2,8 +2,9 @@
   "semi": true,
   "trailingComma": "es5",
   "singleQuote": true,
-  "printWidth": 100,
   "tabWidth": 2,
   "useTabs": false,
-  "endOfLine": "auto"
-} 
+  "printWidth": 100,
+  "bracketSpacing": true,
+  "arrowParens": "always"
+}
